@@ -22,12 +22,11 @@ dynamic = [
 ]
 description = "AsyncIO Service-based programming"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 keywords = ["asyncio", "service", "bootsteps", "graph", "coroutine"]
 authors = [
     { name = "Ask Solem Hoel", email= "ask@robinhood.com" },
 ]
-
 maintainers = [
     { name = "William Barnhart", email = "williambbarnhart@gmail.com" },
 ]
@@ -36,11 +35,11 @@ classifiers = [
     "Development Status :: 5 - Production/Stable",
     "License :: OSI Approved :: BSD License",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
     "Operating System :: POSIX",
@@ -53,7 +52,7 @@ classifiers = [
 ]
 dependencies = [
     "colorlog>=6.0.0,<7.0.0",
-    "croniter>=2.0.0,<3.0.0",
+    "croniter>=2.0.0,<6.0.0",
     "mypy_extensions",
 ]
 
@@ -150,7 +149,7 @@ warn_unused_configs = true
 warn_unused_ignores = true
 
 [tool.ruff]
-target-version = "py38"
+target-version = "py39"
 line-length = 79
 exclude = [
     ".bzr",
