@@ -4,7 +4,7 @@ import math
 import signal
 import traceback
 from types import FrameType
-from typing import Any, Type
+from typing import Any
 
 from .services import Service
 from .utils.logging import get_logger
@@ -59,11 +59,11 @@ class BlockingDetector(Service):
     def __init__(
         self,
         timeout: Seconds,
-        raises: Type[BaseException] = Blocking,
+        raises: type[BaseException] = Blocking,
         **kwargs: Any,
     ) -> None:
         self.timeout: float = want_seconds(timeout)
-        self.raises: Type[BaseException] = raises
+        self.raises: type[BaseException] = raises
         super().__init__(**kwargs)
 
     @Service.task
