@@ -2,19 +2,9 @@
 
 import collections.abc
 import sys
+from collections.abc import AsyncIterable, AsyncIterator, Iterable, Iterator
 from functools import singledispatch
-from typing import (
-    Any,
-    AsyncIterable,
-    AsyncIterator,
-    Iterable,
-    Iterator,
-    List,
-    Optional,
-    Tuple,
-    TypeVar,
-    cast,
-)
+from typing import Any, Optional, TypeVar, cast
 
 __all__ = [
     "aenumerate",
@@ -31,7 +21,7 @@ T = TypeVar("T")
 
 async def aenumerate(
     it: AsyncIterable[T], start: int = 0
-) -> AsyncIterator[Tuple[int, T]]:
+) -> AsyncIterator[tuple[int, T]]:
     """``async for`` version of ``enumerate``."""
     i = start
     async for item in it:
@@ -136,7 +126,7 @@ class arange(AsyncIterable[int]):
         return _ARangeIterator(self, iter(self._range))
 
 
-async def alist(ait: AsyncIterator[T]) -> List[T]:
+async def alist(ait: AsyncIterator[T]) -> list[T]:
     """Convert async generator to list."""
     return [x async for x in ait]
 
@@ -158,7 +148,7 @@ async def aslice(ait: AsyncIterator[T], *slice_args: int) -> AsyncIterator[T]:
         return
 
 
-async def chunks(it: AsyncIterable[T], n: int) -> AsyncIterable[List[T]]:
+async def chunks(it: AsyncIterable[T], n: int) -> AsyncIterable[list[T]]:
     """Split an async iterator into chunks with `n` elements each.
 
     Example:
